changed kernel data types from uint to ulong

+++ src/CircusTent/Impl/CT_OPENCL/CT_OPENCL_KERNELS.cl
@@ -5,36 +5,36 @@
 
 // FIXME:
 __kernel void RAND_ADD(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 )
 {
-    uint i = 0;
-    uint start = 0;
+    ulong i = 0;
+    ulong start = 0;
     #pragma ocl parallel private(start, i)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for (i=start; i<(start+iters); i++) {
-            __atomic_fetch_add( &ARRAY[IDX[i]], (uint)(0x1), __ATOMIC_RELAXED );
+            __atomic_fetch_add( &ARRAY[IDX[i]], (ulong)(0x1), __ATOMIC_RELAXED );
         }
     }
 }
 
 // FIXME:
 __kernel void RAND_CAS(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
-    uint i      = 0;
-    uint start  = 0;
+    ulong i      = 0;
+    ulong start  = 0;
 
     #pragma ocl parallel private(start,i)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for( i=start; i<(start+iters); i++ ){
         __atomic_compare_exchange_n( &ARRAY[IDX[i]], &ARRAY[IDX[i]], ARRAY[IDX[i]],
                                     0, __ATOMIC_RELAXED, __ATOMIC_RELAXED );
@@ -43,37 +43,37 @@ __kernel void RAND_CAS(
 }
 
 __kernel void STRIDE1_ADD(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
+    ulong i      = 0;
+    ulong start  = 0;
 
     #pragma ocl parallel private(start,i)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for( i=start; i<(start+iters); i++ ){
-            __atomic_fetch_add( &ARRAY[i], (uint)(0xF), __ATOMIC_RELAXED );
+            __atomic_fetch_add( &ARRAY[i], (ulong)(0xF), __ATOMIC_RELAXED );
         }
     }
 }
 
 __kernel void STRIDE1_CAS(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
+    ulong i      = 0;
+    ulong start  = 0;
 
     #pragma ocl parallel private(start,i)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for( i=start; i<(start+iters); i++ ){
             __atomic_compare_exchange_n( &ARRAY[i], &ARRAY[i], ARRAY[i],
                                     0, __ATOMIC_RELAXED, __ATOMIC_RELAXED );
@@ -82,39 +82,39 @@ __kernel void STRIDE1_CAS(
 }
 
 __kernel void STRIDEN_ADD(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes,
-        uint stride
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes,
+        ulong stride
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
+    ulong i      = 0;
+    ulong start  = 0;
 
     #pragma ocl parallel private(start,i)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for( i=start; i<(start+iters); i+=stride ){
-        __atomic_fetch_add( &ARRAY[i], (uint)(0xF), __ATOMIC_RELAXED );
+        __atomic_fetch_add( &ARRAY[i], (ulong)(0xF), __ATOMIC_RELAXED );
         }
     }
 }
 
 __kernel void STRIDEN_CAS(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes,
-        uint stride
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes,
+        ulong stride
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
+    ulong i      = 0;
+    ulong start  = 0;
 
     #pragma ocl parallel private(start,i)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for( i=start; i<(start+iters); i+=stride ){
         __atomic_compare_exchange_n( &ARRAY[i], &ARRAY[i], ARRAY[i],
                                     0, __ATOMIC_RELAXED, __ATOMIC_RELAXED );
@@ -123,39 +123,39 @@ __kernel void STRIDEN_CAS(
 }
 
 __kernel void PTRCHASE_ADD(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
+    ulong i      = 0;
+    ulong start  = 0;
 
     #pragma ocl parallel private(start,i)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for( i=0; i<iters; i++ ){
         start = __atomic_fetch_add( &IDX[start],
-                                    (uint)(0x00ull),
+                                    (ulong)(0x00ull),
                                     __ATOMIC_RELAXED );
         }
     }
 }
 
 __kernel void PTRCHASE_CAS(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
+    ulong i      = 0;
+    ulong start  = 0;
 
     #pragma ocl parallel private(start,i)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for( i=0; i<iters; i++ ){
         __atomic_compare_exchange_n( &IDX[start], &start, IDX[start],
                                     0, __ATOMIC_RELAXED, __ATOMIC_RELAXED );
@@ -164,46 +164,46 @@ __kernel void PTRCHASE_CAS(
 }
 
 __kernel void SG_ADD(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
-    uint src    = 0;
-    uint dest   = 0;
-    uint val    = 0;
+    ulong i      = 0;
+    ulong start  = 0;
+    ulong src    = 0;
+    ulong dest   = 0;
+    ulong val    = 0;
 
     #pragma ocl parallel private(start,i,src,dest,val)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for( i=start; i<(start+iters); i++ ){
-        src  = __atomic_fetch_add( &IDX[i], (uint)(0x00ull), __ATOMIC_RELAXED );
-        dest = __atomic_fetch_add( &IDX[i+1], (uint)(0x00ull), __ATOMIC_RELAXED );
-        val = __atomic_fetch_add( &ARRAY[src], (uint)(0x01ull), __ATOMIC_RELAXED );
+        src  = __atomic_fetch_add( &IDX[i], (ulong)(0x00ull), __ATOMIC_RELAXED );
+        dest = __atomic_fetch_add( &IDX[i+1], (ulong)(0x00ull), __ATOMIC_RELAXED );
+        val = __atomic_fetch_add( &ARRAY[src], (ulong)(0x01ull), __ATOMIC_RELAXED );
         __atomic_fetch_add( &ARRAY[dest], val, __ATOMIC_RELAXED );
         }
     }
 }
 
 __kernel void SG_CAS(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
-    uint src    = 0;
-    uint dest   = 0;
-    uint val    = 0;
+    ulong i      = 0;
+    ulong start  = 0;
+    ulong src    = 0;
+    ulong dest   = 0;
+    ulong val    = 0;
 
     #pragma ocl parallel private(start,i,src,dest,val)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         val   = 0x00ull;
         src   = 0x00ull;
         dest  = 0x00ull;
@@ -221,30 +221,30 @@ __kernel void SG_CAS(
 }
 
 __kernel void CENTRAL_ADD(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
+    ulong i      = 0;
 
     #pragma ocl parallel private(i)
     {
         for( i=0; i<iters; i++ ){
-        __atomic_fetch_add( &ARRAY[0], (uint)(0x1), __ATOMIC_RELAXED );
+        __atomic_fetch_add( &ARRAY[0], (ulong)(0x1), __ATOMIC_RELAXED );
         }
     }
 }
 
 __kernel void CENTRAL_CAS(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
+    ulong i      = 0;
 
     #pragma ocl parallel private(i)
     {
@@ -256,43 +256,43 @@ __kernel void CENTRAL_CAS(
 }
 
 __kernel void SCATTER_ADD(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
-    uint dest   = 0;
-    uint val    = 0;
+    ulong i      = 0;
+    ulong start  = 0;
+    ulong dest   = 0;
+    ulong val    = 0;
 
     #pragma ocl parallel private(start,i,dest,val)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for( i=start; i<(start+iters); i++ ){
-        dest = __atomic_fetch_add( &IDX[i+1], (uint)(0x00ull), __ATOMIC_RELAXED );
-        val = __atomic_fetch_add( &ARRAY[i], (uint)(0x01ull), __ATOMIC_RELAXED );
+        dest = __atomic_fetch_add( &IDX[i+1], (ulong)(0x00ull), __ATOMIC_RELAXED );
+        val = __atomic_fetch_add( &ARRAY[i], (ulong)(0x01ull), __ATOMIC_RELAXED );
         __atomic_fetch_add( &ARRAY[dest], val, __ATOMIC_RELAXED );
         }
     }
 }
 
 __kernel void SCATTER_CAS(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
-    uint dest   = 0;
-    uint val    = 0;
+    ulong i      = 0;
+    ulong start  = 0;
+    ulong dest   = 0;
+    ulong val    = 0;
 
     #pragma ocl parallel private(start,i,dest,val)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         dest  = 0x00ull;
         val   = 0x00ull;
         for( i=start; i<(start+iters); i++ ){
@@ -307,43 +307,43 @@ __kernel void SCATTER_CAS(
 }
 
 __kernel void GATHER_ADD(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
-    uint dest   = 0;
-    uint val    = 0;
+    ulong i      = 0;
+    ulong start  = 0;
+    ulong dest   = 0;
+    ulong val    = 0;
 
     #pragma ocl parallel private(start,i,dest,val)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         for( i=start; i<(start+iters); i++ ){
-        dest = __atomic_fetch_add( &IDX[i+1], (uint)(0x00ull), __ATOMIC_RELAXED );
-        val = __atomic_fetch_add( &ARRAY[dest], (uint)(0x01ull), __ATOMIC_RELAXED );
+        dest = __atomic_fetch_add( &IDX[i+1], (ulong)(0x00ull), __ATOMIC_RELAXED );
+        val = __atomic_fetch_add( &ARRAY[dest], (ulong)(0x01ull), __ATOMIC_RELAXED );
         __atomic_fetch_add( &ARRAY[i], val, __ATOMIC_RELAXED );
         }
     }
 }
 
 __kernel void GATHER_CAS(
-        __local uint* ARRAY,
-        __local uint* IDX,
-        uint iters,
-        uint pes
+        __local ulong* ARRAY,
+        __local ulong* IDX,
+        ulong iters,
+        ulong pes
 ) {
     // todo
-    uint i      = 0;
-    uint start  = 0;
-    uint dest   = 0;
-    uint val    = 0;
+    ulong i      = 0;
+    ulong start  = 0;
+    ulong dest   = 0;
+    ulong val    = 0;
 
     #pragma ocl parallel private(start,i,dest,val)
     {
-        start = (uint)(get_global_id(0)) * iters;
+        start = (ulong)(get_global_id(0)) * iters;
         dest  = 0x00ull;
         val   = 0x00ull;
         for( i=start; i<(start+iters); i++ ){
